requirements.txt
# Raspberry Pi 4 - Nutriquant Bağımlılıkları

# Kamera
picamera2==0.3.12
libcamera

# Görüntü İşleme ve AI
tflite-runtime==2.14.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3

# Tartı Sensörü
hx711==1.1.2.3
RPi.GPIO==0.7.1

# UI - Dokunmatik Ekran
pygame==2.5.2
tkinter

# Ses
pygame-mixer
pydub==0.25.1

# UPS HAT - Pil Yönetimi
smbus2==0.4.3
ina219==1.4.0

# LED Kontrol
rpi-ws281x==5.0.0
adafruit-circuitpython-neopixel==6.3.11

# Veritabanı
python-json-logger==2.0.7

# Yardımcı
python-dotenv==1.0.0
